+++ application/src/main/java/common/bolts/transactional/ed/wug/WUGBolt.java
@@ -27,7 +27,7 @@ public abstract class WUGBolt extends TransactionalBolt {
     }
 
     @Override
-    protected void TXN_PROCESS(long _bid) throws DatabaseException, InterruptedException {
+    protected void TXN_PROCESS(double _bid) throws DatabaseException, InterruptedException {
     }
 
     protected void WU_GATE_REQUEST_POST(WUEvent event) throws InterruptedException, DatabaseException {
