+++ state-engine/src/main/java/scheduler/impl/og/structured/OGDFSAScheduler.java
@@ -159,7 +159,7 @@ public class OGDFSAScheduler extends AbstractOGDFSScheduler<OGSAContext> {
      * @return
      */
     private boolean _MarkOperationsToAbort(OGSAContext context, Operation operation) {
-        long bid = operation.bid;
+        double bid = operation.bid;
         boolean markAny = false;
         //identify bids to be aborted.
         for (Operation failedOp : failedOperations) {
